pyproject.toml
[project]
name = "tourism-monitor"
version = "0.1.0"
description = "Tourism Monitor — Cozumel Cruise Data Pipeline & Dashboard"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.43.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "statsmodels>=0.14.0",
    "scikit-learn>=1.4.0",
    "scipy>=1.13.0",
    "plotly>=5.20.0",
    "pydeck>=0.9.0",
    "streamlit>=1.35.0",
    "supabase>=2.4.0,<2.10.0",
    "openmeteo-requests>=1.2.0",
    "requests-cache>=1.2.0",
    "retry-requests>=2.0.0",
    "rapidfuzz>=3.8.0",
    "python-dotenv>=1.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.29.0",
    "jupyter>=1.0.0",
]
